+++ Assets/UGF.Module.Descriptions.Runtime.Tests/New Description Module Asset.asset
@@ -14,14 +14,19 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   m_descriptions:
   - m_guid:
-      m_first: 4675955830162030254
-      m_second: 3654966563721544
+      serializedVersion: 2
+      Hash: 00000000000000000000000000000000
     m_asset: {fileID: 11400000, guid: a2397eae550d40e44851b2372cfc0c00, type: 2}
   m_collections:
   - m_guid:
-      m_first: 2541164659296716583
-      m_second: 1025253101817878728
+      serializedVersion: 2
+      Hash: 00000000000000000000000000000000
     m_asset: {fileID: 11400000, guid: 4128472707cb2344c8d8b89a436e3a0e, type: 2}
+  m_tables:
+  - m_guid:
+      serializedVersion: 2
+      Hash: 7902574e625f0ee4b9dc95b2dd106979
+    m_asset: {fileID: 11400000, guid: 7902574e625f0ee4b9dc95b2dd106979, type: 2}
   m_loadAsync:
-  - m_first: 4675955830162030254
-    m_second: 3654966563721544
+  - serializedVersion: 2
+    Hash: 00000000000000000000000000000000
