--- Assets/UGF.Module.Descriptions.Runtime.Tests/New Test Table Asset.asset
@@ -1,36 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 65d7338b4d9e4841a1f498e5b721ff3e, type: 3}
-  m_Name: New Test Table Asset
-  m_EditorClassIdentifier: 
-  m_table:
-    m_entries:
-    - m_id:
-        serializedVersion: 2
-        Hash: 3a415b20752f4239a577cc4f835399fb
-      m_name: Entry
-      m_value: 0
-    - m_id:
-        serializedVersion: 2
-        Hash: 3407c297acec420fa89458c35d7b396a
-      m_name: Entry (1)
-      m_value: 0
-    - m_id:
-        serializedVersion: 2
-        Hash: 605fa5506cf24b9dba497e30038f9451
-      m_name: Entry (2)
-      m_value: 0
-    - m_id:
-        serializedVersion: 2
-        Hash: e847bdd449374d57b8f53b7fb2a6baff
-      m_name: Entry (3)
-      m_value: 0
