--- Assets/UGF.Module.Descriptions.Runtime.Tests/New Test Table Asset.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dec3391db722f724ca0be1512a4dfefa
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
