--- Assets/UGF.Module.Descriptions.Runtime.Tests/TestTableAsset.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 65d7338b4d9e4841a1f498e5b721ff3e
-timeCreated: 1723745359
